Added the bcc instruction and added a small bugfix

+++ src/arm64/helper.rs
@@ -1,9 +1,9 @@
 use tnj::types::{Type, I32, I64};
 use yaxpeax_arm::armv8::a64::{Operand, SizeCode};
 
-pub fn get_pc_offset_as_int(operand: Operand) -> u64 {
+pub fn get_pc_offset_as_int(operand: Operand) -> i64 {
     match operand {
-        Operand::PCOffset(imm) => imm as u64,
+        Operand::PCOffset(imm) => imm,
         op => unimplemented!("dst op {:?}", op),
     }
 }
